+++ app/controllers/recording_studio_webhooks/admin/action_attempts_controller.rb
@@ -2,13 +2,12 @@
 
 module RecordingStudioWebhooks
   module Admin
-    class ActionPlansController < BaseController
+    class ActionAttemptsController < BaseController
       before_action :load_endpoint, if: -> { params[:endpoint_id].present? }
       before_action :load_event, if: -> { params[:event_id].present? }
-      before_action :load_action_plan
+      before_action :load_action_attempt
 
-      def show
-      end
+      def show; end
 
       private
 
@@ -18,13 +17,13 @@ module RecordingStudioWebhooks
         raise ActionController::RoutingError, "Not Found"
       end
 
-      def load_action_plan
-        @action_plan = if @event
-                         @event.action_plans.find(params[:id])
-                       else
-                         ActionPlan.includes(inbound_event: :endpoint).find(params[:id])
-                       end
-        @event ||= @action_plan.inbound_event
+      def load_action_attempt
+        @action_attempt = if @event
+                            @event.action_attempts.find(params[:id])
+                          else
+                            ActionAttempt.includes(inbound_event: :endpoint).find(params[:id])
+                          end
+        @event ||= @action_attempt.inbound_event
         @endpoint ||= @event.endpoint
       rescue ActiveRecord::RecordNotFound
         raise ActionController::RoutingError, "Not Found"