+++ app/jobs/recording_studio_webhooks/execute_action_plan_active_job.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module RecordingStudioWebhooks
-  class ExecuteActionPlanActiveJob < ActiveJob::Base
-    def perform(action_plan_id)
-      ExecuteActionPlan.call(action_plan_id)
-    end
-  end
-end